+++ CavyLifeBand2/AccountInfo/ViewModel/AccountSettingViewModel.swift
@@ -6,8 +6,9 @@
 //  Copyright © 2016年 xuemincai. All rights reserved.
 //
 import UIKit
+import JSONJoy
 import RealmSwift
-
+import EZSwiftExtensions
 
 typealias AccountSettingModelPotocols = protocol<GuideViewDataSource, GuideViewDelegate, SetUserInfoRequestsDelegate, UserInfoRealmOperateDelegate>
 
@@ -167,3 +168,125 @@ struct AccountWeightViewModel: AccountSettingModelPotocols {
     }
     
 }
+
+/**
+ *  登录用户修改昵称
+ */
+class UserChangeNicknameVM: ContactsReqFriendPortocols, SetUserInfoRequestsDelegate, UserInfoRealmOperateDelegate {
+    
+    var realm: Realm = try! Realm()
+    
+    var navTitle: String = L10n.AccountInfoTitle.string
+    
+    var textFieldTitle: String {
+        didSet {
+            userInfoPara[UserNetRequsetKey.NickName.rawValue] = textFieldTitle
+        }
+    }
+    
+    var placeholderText: String {
+        return L10n.AccountInofChangeNicknamePlaceholder.string
+    }
+    
+    var bottonTitle: String {
+        return L10n.ContactsRequestSureButton.string
+    }
+    
+    var friendId: String = ""
+    
+    weak var viewController: UIViewController?
+    
+    var userInfoPara: [String: AnyObject] = [UserNetRequsetKey.UserID.rawValue: CavyDefine.loginUserBaseInfo.loginUserInfo.loginUserId]
+    
+    //点击发送请求成功回调
+    var onClickButtonCellBack: (String -> Void)?
+    
+    init(viewController: UIViewController, onClickButtonCellBack: (String -> Void)? = nil) {
+        
+        self.viewController = viewController
+        self.onClickButtonCellBack = onClickButtonCellBack
+        self.textFieldTitle = ""
+    }
+    
+    func onClickButton() {
+        
+        setUserInfo { [unowned self] in
+            
+            if $0 {
+                
+                self.updateUserInfo(CavyDefine.loginUserBaseInfo.loginUserInfo.loginUserId) { [unowned self] in
+                    $0.nickname = self.textFieldTitle
+                    return $0
+                }
+                
+                self.viewController?.popVC()
+            }
+            
+        }
+        
+    }
+    
+    
+}
+
+/**
+ *  登录用户修改地址
+ */
+class UserChangeAddressVM: ContactsReqFriendPortocols, SetUserInfoRequestsDelegate, UserInfoRealmOperateDelegate {
+    
+    var realm: Realm = try! Realm()
+    
+    var navTitle: String = L10n.AccountInfoTitle.string
+    
+    var textFieldTitle: String {
+        didSet {
+            userInfoPara[UserNetRequsetKey.Address.rawValue] = textFieldTitle
+        }
+    }
+    
+    var placeholderText: String {
+        return L10n.AccountInofChangeAddressPlaceholder.string
+    }
+    
+    var bottonTitle: String {
+        return L10n.ContactsRequestSureButton.string
+    }
+    
+    var friendId: String = ""
+    
+    weak var viewController: UIViewController?
+    
+    var userInfoPara: [String: AnyObject] = [UserNetRequsetKey.UserID.rawValue: CavyDefine.loginUserBaseInfo.loginUserInfo.loginUserId]
+    
+    //点击发送请求成功回调
+    var onClickButtonCellBack: (String -> Void)?
+    
+    init(viewController: UIViewController, onClickButtonCellBack: (String -> Void)? = nil) {
+        
+        self.viewController = viewController
+        self.onClickButtonCellBack = onClickButtonCellBack
+        self.textFieldTitle = ""
+    }
+    
+    func onClickButton() {
+        
+        setUserInfo { [unowned self] in
+            
+            if $0 {
+                
+                self.updateUserInfo(CavyDefine.loginUserBaseInfo.loginUserInfo.loginUserId) { [unowned self] in
+                    $0.address = self.textFieldTitle
+                    return $0
+                }
+                
+                self.viewController?.popVC()
+            }
+            
+        }
+        
+    }
+    
+    
+}
+
+
